Move eyaml to a new library in `shared/`. (#1216)
Moves the event parser as well, which is a non-YAML-compliant dialect of a YAML-like language.
Creates a general format for shared code and includes it from the CompilerSource makefile.
Handles an aggravating problem with `wildcard` in an included makefile.
