+++ CompilerSource/Makefile
@@ -1,7 +1,7 @@
-# Awesomer Makefile generated by Rusky: stop relying on Bash hacks and automate more stuff
-# It will automatically build the shared library out of any .cpp files in this directory or its
-# subdirectories (recursively). It will also automatically calculate and update all dependencies
-# automatically when 'make' is run, so there is no need to run genmake.sh EVER.
+# "Shared" includes.
+include ../shared/Makefile
+include ../shared/eyaml/Makefile
+include ../shared/event_reader/Makefile
 
 #################
 # configuration #
@@ -39,6 +39,11 @@ SOURCES := $(filter-out ./standalone_main.cpp, $(call rwildcard,./,*.cpp))
 OBJECTS := $(addprefix .eobjs/,$(SOURCES:.cpp=.o))
 DEPENDS := $(OBJECTS:.o=.d)
 
+CXXFLAGS += -I../shared $(addprefix -I../shared/, $(SHARED_INCLUDES))
+SOURCES += $(addprefix ../shared/,$(SHARED_SOURCES))
+OBJECTS += $(addprefix .eobjs/shared/,$(SHARED_SOURCES:.cpp=.o))
+DEPENDS += $(addprefix .eobjs/shared/,$(SHARED_SOURCES:.cpp=.d))
+
 # sort has the nice side effect of removing duplicates. this may or may not be faster.
 OBJDIRS := $(sort $(dir $(OBJECTS)))
 override CPPFLAGS += -I.
@@ -62,6 +67,8 @@ $(TARGET): $(OBJECTS)
 # -MP gives phony rules for non-target files, avoiding problems with missing files
 .eobjs/%.o .eobjs/%.d: %.cpp | $(OBJDIRS)
 	$(CXX) $(CXXFLAGS) $(CPPFLAGS) -MMD -MP -c -o .eobjs/$*.o $<
+.eobjs/shared/%.o .eobjs/shared/%.d: ../shared/%.cpp | $(OBJDIRS)
+	$(CXX) $(CXXFLAGS) $(CPPFLAGS) -MMD -MP -c -o .eobjs/shared/$*.o $<
 
 $(OBJDIRS):
 	$(MKDIR) -p $@
