+++ CompilerSource/compiler/components/parse_and_link.cpp
@@ -39,7 +39,7 @@ using namespace std;
 #include "backend/EnigmaStruct.h" //LateralGM interface structures
 #include "parser/object_storage.h"
 #include "compiler/compile_common.h"
-#include "compiler/event_reader/event_parser.h"
+#include "event_reader/event_parser.h"
 
 #include <math.h> //log2 to calculate passes.
