+++ CompilerSource/compiler/components/write_defragged_events.cpp
@@ -34,7 +34,7 @@ using namespace std;
 #include "backend/EnigmaStruct.h" //LateralGM interface structures
 #include "compiler/compile_common.h"
 
-#include "compiler/event_reader/event_parser.h"
+#include "event_reader/event_parser.h"
 #include "languages/lang_CPP.h"
 
 struct foundevent { int mid, id, count; foundevent(): mid(0),id(0),count(0) {} void f2(int m,int i) { id = i, mid = m; } void inc(int m,int i) { mid=m,id=i,count++; } void operator++(int) { count++; } };
