+++ CompilerSource/compiler/components/write_object_data.cpp
@@ -30,7 +30,7 @@ using namespace std;
 
 #include "backend/EnigmaStruct.h" //LateralGM interface structures
 #include "compiler/compile_common.h"
-#include "compiler/event_reader/event_parser.h"
+#include "event_reader/event_parser.h"
 #include "general/parse_basics_old.h"
 #include "settings.h"
