+++ shared/event_reader/event_parser.cpp
