+++ shared/eyaml/Makefile
@@ -1,0 +1,2 @@
+# SHARED_INCLUDES += eyaml (Best to just let them include eyaml/eyaml.h)
+SHARED_SOURCES := $(SHARED_SOURCES) $(addprefix eyaml/,$(call relative_wildcard,*.cpp))
