+++ shared/eyaml/eyaml.cpp
